rewrote deploy script to also copy folders in the .config path

+++ deploy.sh
@@ -4,15 +4,40 @@
 path=$(pwd)
 
 # Update symlinks in home
-for link in `ls | grep -vE "README\.md|deploy\.sh"`; do
+for link in `ls | grep -vE "README\.md|deploy\.sh|config"`; do
   [ "x$verbose" == 'x' ] || echo "> ${link}"
 
-  [ "x$verbose" == 'x' ] echo "  removing ~/.${link}"
-  rm ~/.${link}
+  # remove symlink if exists
+  if [ -e ~/.${link} ]; then 
+    [ "x$verbose" == 'x' ] || echo "  removing ~/.${link}"
+    if [ -h ~/.${link} ]; then
+      rm ~/.${link}
+    else
+      echo "~/.${link} exists and is not an symlink!"
+      exit 1
+    fi
+  fi
 
   [ "x$verbose" == 'x' ] || echo "  creating new symlink: ${path}/${link} -> ~/.${link}"
   ln -s ${path}/${link} ~/.${link}
+done
 
+for link in `ls config`; do
+  [ "x$verbose" == 'x' ] || echo "> ${link}"
+
+  # remove symlink if exists
+  if [ -e ~/.config/${link} ]; then 
+    [ "x$verbose" == 'x' ] || echo "  removing ~/.config/${link}"
+    if [ -h ~/.config/${link} ]; then
+      rm ~/.config/${link}
+    else
+      echo "~/.config/${link} exists and is not an symlink!"
+      exit 1
+    fi
+  fi
+
+  [ "x$verbose" == 'x' ] || echo "  creating new symlink: ${path}/config/${link} -> ~/.config/${link}"
+  ln -s ${path}/config/${link} ~/.config/${link}
 done
 
 [ -z "$verbose" ] && echo "Files deployed"
